==> 1_Limpieza.ipynb <==
## 1. Datasets

import pandas as pd
import numpy


df_meta = pd.read_csv('./The_Movies_Dataset/movies_metadata.csv', low_memory=False)
df_ratings = pd.read_csv('./The_Movies_Dataset/ratings_small.csv', low_memory=False)


df_meta.head(3)

df_ratings.head()

## 2. Limpieza

import pandas as pd
import numpy as np
import ast # Necesario para evaluar las cadenas JSON en 'genres'

# Configuración de Pandas para mejor visualización
pd.set_option('display.max_columns', None)

## ----------- 1: LIMPIEZA DE movies_metadata -------------------------

# low_memory=False se usa para evitar advertencias de pandas con datasets grandes y heterogéneos
df_meta = pd.read_csv('./The_Movies_Dataset/movies_metadata.csv', low_memory=False)

# 1. Limpieza de IDs 
df_meta = df_meta[df_meta['id'].str.isnumeric()]

df_meta['id'] = df_meta['id'].astype(int)
print(f"Filas después de limpiar IDs: {len(df_meta):,}")

# 2. Valores Faltantes  en 'overview'

df_meta['overview'] = df_meta['overview'].fillna('')
print("Vacios en 'overview' rellenados")

# 3. Parsing de la columna 'genres' (JSON)

def extract_names(json_string):
    if pd.isna(json_string) or json_string == '':
        return []
    try:
        # ast.literal_eval es más seguro que eval()
        list_of_dicts = ast.literal_eval(json_string)
        return [d['name'] for d in list_of_dicts]
    except:
        # En caso de error de parsing (malformación), devuelve vacío
        return []

df_meta['genres'] = df_meta['genres'].apply(extract_names)
print("Columna 'genres' parseada a una lista de nombres")

# 4. Selección de Columnas para el Sistema de Recomendación
df_movies_cleaned = df_meta[['id', 'title', 'overview', 'genres', 'vote_average', 'release_date']].copy()
df_movies_cleaned.rename(columns={'id': 'movieId'}, inplace=True) # Renombrar para hacer match con ratings

print(f"movies_metadata limpio: {len(df_movies_cleaned):,} registros\n")




##  ------------2: LIMPIEZA DE ratings ------------------------

df_ratings = pd.read_csv('./The_Movies_Dataset/ratings_small.csv')

df_ratings['userId'] = df_ratings['userId'].astype(int)
df_ratings['movieId'] = df_ratings['movieId'].astype(int)

print(f"NaNs en ratings:\n{df_ratings.isnull().sum()}")

print(f"\nRegistros {len(df_ratings):,}")
print(f"Usuarios únicos: {df_ratings['userId'].nunique():,} \nPelículas únicas en ratings: {df_ratings['movieId'].nunique():,}\n")


## --------------3: Merge de datasets ---------------------

# Hacemos merge (solo películas con ratings) para garantizar IDs válidos
df_merged = pd.merge(df_ratings, df_movies_cleaned, on='movieId', how='inner')
df_merged=df_merged.drop(['timestamp','release_date'], axis=1)

print("--- Fusión de Datasets (solo películas con ratings) ---")
print(f"Registros después del merge (ratings válidos con metadatos): {len(df_merged):,}")
print(f"Columnas resultantes: {df_merged.columns.tolist()}")

df_merged.head()

df_merged.to_csv('./archivos/dataset_merged.csv', index=False)
# df_movies_cleaned.to_csv('cleaned_movie_content.csv', index=False)
